rename util/interfaces to util/types
