+++ docs/interfaces/_staking_stakingtxtypeutils_.txinfounbond.md
@@ -4,7 +4,7 @@
 
 ## Hierarchy
 
-* [BaseTxInfo](_util_interfaces_.basetxinfo.md)
+* [BaseTxInfo](_util_types_.basetxinfo.md)
 
   ↳ **TxInfoUnbond**
 
@@ -29,9 +29,9 @@
 
 • **address**: *string*
 
-*Inherited from [BaseTxInfo](_util_interfaces_.basetxinfo.md).[address](_util_interfaces_.basetxinfo.md#address)*
+*Inherited from [BaseTxInfo](_util_types_.basetxinfo.md).[address](_util_types_.basetxinfo.md#address)*
 
-*Defined in [src/util/interfaces.ts:21](https://github.com/paritytech/txwrapper/blob/0fb6503/src/util/interfaces.ts#L21)*
+*Defined in [src/util/types.ts:21](https://github.com/paritytech/txwrapper/blob/fcbe6db/src/util/types.ts#L21)*
 
 The ss-58 encoded address of the sending account.
 
@@ -41,9 +41,9 @@ ___
 
 • **blockHash**: *string*
 
-*Inherited from [BaseTxInfo](_util_interfaces_.basetxinfo.md).[blockHash](_util_interfaces_.basetxinfo.md#blockhash)*
+*Inherited from [BaseTxInfo](_util_types_.basetxinfo.md).[blockHash](_util_types_.basetxinfo.md#blockhash)*
 
-*Defined in [src/util/interfaces.ts:25](https://github.com/paritytech/txwrapper/blob/0fb6503/src/util/interfaces.ts#L25)*
+*Defined in [src/util/types.ts:25](https://github.com/paritytech/txwrapper/blob/fcbe6db/src/util/types.ts#L25)*
 
 The checkpoint hash of the block, in hex.
 
@@ -53,9 +53,9 @@ ___
 
 • **blockNumber**: *number*
 
-*Inherited from [BaseTxInfo](_util_interfaces_.basetxinfo.md).[blockNumber](_util_interfaces_.basetxinfo.md#blocknumber)*
+*Inherited from [BaseTxInfo](_util_types_.basetxinfo.md).[blockNumber](_util_types_.basetxinfo.md#blocknumber)*
 
-*Defined in [src/util/interfaces.ts:29](https://github.com/paritytech/txwrapper/blob/0fb6503/src/util/interfaces.ts#L29)*
+*Defined in [src/util/types.ts:29](https://github.com/paritytech/txwrapper/blob/fcbe6db/src/util/types.ts#L29)*
 
 The checkpoint block number (u32), in hex.
 
@@ -65,9 +65,9 @@ ___
 
 • **genesisHash**: *string*
 
-*Inherited from [BaseTxInfo](_util_interfaces_.basetxinfo.md).[genesisHash](_util_interfaces_.basetxinfo.md#genesishash)*
+*Inherited from [BaseTxInfo](_util_types_.basetxinfo.md).[genesisHash](_util_types_.basetxinfo.md#genesishash)*
 
-*Defined in [src/util/interfaces.ts:33](https://github.com/paritytech/txwrapper/blob/0fb6503/src/util/interfaces.ts#L33)*
+*Defined in [src/util/types.ts:33](https://github.com/paritytech/txwrapper/blob/fcbe6db/src/util/types.ts#L33)*
 
 The genesis hash of the chain, in hex.
 
@@ -77,9 +77,9 @@ ___
 
 • **metadataRpc**: *string*
 
-*Inherited from [BaseTxInfo](_util_interfaces_.basetxinfo.md).[metadataRpc](_util_interfaces_.basetxinfo.md#metadatarpc)*
+*Inherited from [BaseTxInfo](_util_types_.basetxinfo.md).[metadataRpc](_util_types_.basetxinfo.md#metadatarpc)*
 
-*Defined in [src/util/interfaces.ts:38](https://github.com/paritytech/txwrapper/blob/0fb6503/src/util/interfaces.ts#L38)*
+*Defined in [src/util/types.ts:38](https://github.com/paritytech/txwrapper/blob/fcbe6db/src/util/types.ts#L38)*
 
 The SCALE-encoded metadata, as a hex string. Can be retrieved via the RPC
 call `state_getMetadata`.
@@ -90,9 +90,9 @@ ___
 
 • **nonce**: *number*
 
-*Inherited from [BaseTxInfo](_util_interfaces_.basetxinfo.md).[nonce](_util_interfaces_.basetxinfo.md#nonce)*
+*Inherited from [BaseTxInfo](_util_types_.basetxinfo.md).[nonce](_util_types_.basetxinfo.md#nonce)*
 
-*Defined in [src/util/interfaces.ts:42](https://github.com/paritytech/txwrapper/blob/0fb6503/src/util/interfaces.ts#L42)*
+*Defined in [src/util/types.ts:42](https://github.com/paritytech/txwrapper/blob/fcbe6db/src/util/types.ts#L42)*
 
 The nonce for this transaction.
 
@@ -102,9 +102,9 @@ ___
 
 • **specVersion**: *number*
 
-*Inherited from [BaseTxInfo](_util_interfaces_.basetxinfo.md).[specVersion](_util_interfaces_.basetxinfo.md#specversion)*
+*Inherited from [BaseTxInfo](_util_types_.basetxinfo.md).[specVersion](_util_types_.basetxinfo.md#specversion)*
 
-*Defined in [src/util/interfaces.ts:46](https://github.com/paritytech/txwrapper/blob/0fb6503/src/util/interfaces.ts#L46)*
+*Defined in [src/util/types.ts:46](https://github.com/paritytech/txwrapper/blob/fcbe6db/src/util/types.ts#L46)*
 
 The current spec version for the runtime.
 
@@ -114,9 +114,9 @@ ___
 
 • **tip**: *number*
 
-*Inherited from [BaseTxInfo](_util_interfaces_.basetxinfo.md).[tip](_util_interfaces_.basetxinfo.md#tip)*
+*Inherited from [BaseTxInfo](_util_types_.basetxinfo.md).[tip](_util_types_.basetxinfo.md#tip)*
 
-*Defined in [src/util/interfaces.ts:50](https://github.com/paritytech/txwrapper/blob/0fb6503/src/util/interfaces.ts#L50)*
+*Defined in [src/util/types.ts:50](https://github.com/paritytech/txwrapper/blob/fcbe6db/src/util/types.ts#L50)*
 
 The tip for this transaction, in hex.
 
@@ -126,9 +126,9 @@ ___
 
 • **validityPeriod**: *number*
 
-*Inherited from [BaseTxInfo](_util_interfaces_.basetxinfo.md).[validityPeriod](_util_interfaces_.basetxinfo.md#validityperiod)*
+*Inherited from [BaseTxInfo](_util_types_.basetxinfo.md).[validityPeriod](_util_types_.basetxinfo.md#validityperiod)*
 
-*Defined in [src/util/interfaces.ts:55](https://github.com/paritytech/txwrapper/blob/0fb6503/src/util/interfaces.ts#L55)*
+*Defined in [src/util/types.ts:55](https://github.com/paritytech/txwrapper/blob/fcbe6db/src/util/types.ts#L55)*
 
 The amount of time (in second) the transaction is valid for. Will be
 translated into a mortal era.
@@ -139,6 +139,6 @@ ___
 
 • **value**: *number*
 
-*Defined in [src/staking/stakingTxTypeUtils.ts:32](https://github.com/paritytech/txwrapper/blob/0fb6503/src/staking/stakingTxTypeUtils.ts#L32)*
+*Defined in [src/staking/stakingTxTypeUtils.ts:32](https://github.com/paritytech/txwrapper/blob/fcbe6db/src/staking/stakingTxTypeUtils.ts#L32)*
 
 The number of tokens to unbond.
