+++ docs/modules/_createsignedtx_.md
@@ -12,19 +12,19 @@
 
 ###  createSignedTx
 
-▸ **createSignedTx**(`unsigned`: [UnsignedTransaction](../interfaces/_util_interfaces_.unsignedtransaction.md), `signature`: string): *string*
+▸ **createSignedTx**(`unsigned`: [UnsignedTransaction](../interfaces/_util_types_.unsignedtransaction.md), `signature`: string): *string*
 
-*Defined in [src/createSignedTx.ts:14](https://github.com/paritytech/txwrapper/blob/0fb6503/src/createSignedTx.ts#L14)*
+*Defined in [src/createSignedTx.ts:14](https://github.com/paritytech/txwrapper/blob/fcbe6db/src/createSignedTx.ts#L14)*
 
 Serialize a signed transaction in a format that can be submitted over the
 Node RPC Interface from the signing payload and signature produced by the
-remote signer
+remote signer.
 
 **Parameters:**
 
 Name | Type | Description |
 ------ | ------ | ------ |
-`unsigned` | [UnsignedTransaction](../interfaces/_util_interfaces_.unsignedtransaction.md) | The JSON representing the unsigned transaction |
-`signature` | string | Signature of the signing payload produced by the remote signer  |
+`unsigned` | [UnsignedTransaction](../interfaces/_util_types_.unsignedtransaction.md) | The JSON representing the unsigned transaction. |
+`signature` | string | Signature of the signing payload produced by the remote signer.  |
 
 **Returns:** *string*
