+++ docs/modules/_decode_decode_.md
@@ -12,9 +12,9 @@
 
 ###  decode
 
-▸ **decode**(`unsignedTx`: [UnsignedTransaction](../interfaces/_util_interfaces_.unsignedtransaction.md), `metadataRpc`: string): *DecodedUnsignedTx*
+▸ **decode**(`unsignedTx`: [UnsignedTransaction](../interfaces/_util_types_.unsignedtransaction.md), `metadataRpc`: string): *DecodedUnsignedTx*
 
-*Defined in [src/decode/decode.ts:17](https://github.com/paritytech/txwrapper/blob/0fb6503/src/decode/decode.ts#L17)*
+*Defined in [src/decode/decode.ts:16](https://github.com/paritytech/txwrapper/blob/fcbe6db/src/decode/decode.ts#L16)*
 
 Parse the transaction information from a signing payload, an unsigned tx, or a signed tx.
 
@@ -22,14 +22,14 @@ Parse the transaction information from a signing payload, an unsigned tx, or a s
 
 Name | Type | Description |
 ------ | ------ | ------ |
-`unsignedTx` | [UnsignedTransaction](../interfaces/_util_interfaces_.unsignedtransaction.md) | The data to parse, as an unsigned tx. |
+`unsignedTx` | [UnsignedTransaction](../interfaces/_util_types_.unsignedtransaction.md) | The data to parse, as an unsigned tx. |
 `metadataRpc` | string | The SCALE-encoded metadata, as a hex string. Can be retrieved via the RPC call `state_getMetadata`.  |
 
 **Returns:** *DecodedUnsignedTx*
 
 ▸ **decode**(`signedTx`: string, `metadataRpc`: string): *DecodedSignedTx*
 
-*Defined in [src/decode/decode.ts:29](https://github.com/paritytech/txwrapper/blob/0fb6503/src/decode/decode.ts#L29)*
+*Defined in [src/decode/decode.ts:28](https://github.com/paritytech/txwrapper/blob/fcbe6db/src/decode/decode.ts#L28)*
 
 Parse the transaction information from a signing payload, an unsigned tx, or a signed tx.
 
@@ -44,7 +44,7 @@ Name | Type | Description |
 
 ▸ **decode**(`signingPayload`: string, `metadataRpc`: string): *DecodedSigningPayload*
 
-*Defined in [src/decode/decode.ts:38](https://github.com/paritytech/txwrapper/blob/0fb6503/src/decode/decode.ts#L38)*
+*Defined in [src/decode/decode.ts:37](https://github.com/paritytech/txwrapper/blob/fcbe6db/src/decode/decode.ts#L37)*
 
 Parse the transaction information from a signing payload, an unsigned tx, or a signed tx.
