+++ docs/modules/_deriveaddress_.md
@@ -14,7 +14,7 @@
 
 ▸ **deriveAddress**(`publicKey`: string | Uint8Array, `ss58Format`: number): *string*
 
-*Defined in [src/deriveAddress.ts:11](https://github.com/paritytech/txwrapper/blob/0fb6503/src/deriveAddress.ts#L11)*
+*Defined in [src/deriveAddress.ts:11](https://github.com/paritytech/txwrapper/blob/fcbe6db/src/deriveAddress.ts#L11)*
 
 Derive an address from a cryptographic public key offline
