+++ docs/modules/_staking_bond_.md
@@ -12,9 +12,9 @@
 
 ###  bond
 
-▸ **bond**(`info`: [TxInfoBond](../interfaces/_staking_stakingtxtypeutils_.txinfobond.md)): *[UnsignedTransaction](../interfaces/_util_interfaces_.unsignedtransaction.md)*
+▸ **bond**(`info`: [TxInfoBond](../interfaces/_staking_stakingtxtypeutils_.txinfobond.md)): *[UnsignedTransaction](../interfaces/_util_types_.unsignedtransaction.md)*
 
-*Defined in [src/staking/bond.ts:13](https://github.com/paritytech/txwrapper/blob/0fb6503/src/staking/bond.ts#L13)*
+*Defined in [src/staking/bond.ts:13](https://github.com/paritytech/txwrapper/blob/fcbe6db/src/staking/bond.ts#L13)*
 
 Construct a transaction to bond funds and create a Stash account.
 
@@ -24,4 +24,4 @@ Name | Type | Description |
 ------ | ------ | ------ |
 `info` | [TxInfoBond](../interfaces/_staking_stakingtxtypeutils_.txinfobond.md) | Information required to construct the transaction.  |
 
-**Returns:** *[UnsignedTransaction](../interfaces/_util_interfaces_.unsignedtransaction.md)*
+**Returns:** *[UnsignedTransaction](../interfaces/_util_types_.unsignedtransaction.md)*
