+++ docs/modules/_util_constants_.md
@@ -16,7 +16,7 @@
 
 • **EXTRINSIC_VERSION**: *4* = 4
 
-*Defined in [src/util/constants.ts:13](https://github.com/paritytech/txwrapper/blob/0fb6503/src/util/constants.ts#L13)*
+*Defined in [src/util/constants.ts:13](https://github.com/paritytech/txwrapper/blob/fcbe6db/src/util/constants.ts#L13)*
 
 Latest extrinsic version
 
@@ -26,7 +26,7 @@ ___
 
 • **KUSAMA_SS58_FORMAT**: *2* = 2
 
-*Defined in [src/util/constants.ts:4](https://github.com/paritytech/txwrapper/blob/0fb6503/src/util/constants.ts#L4)*
+*Defined in [src/util/constants.ts:4](https://github.com/paritytech/txwrapper/blob/fcbe6db/src/util/constants.ts#L4)*
 
 Prefix for SS58-encoded addresses on Kusama
 
@@ -36,6 +36,6 @@ ___
 
 • **POLKADOT_SS58_FORMAT**: *0* = 0
 
-*Defined in [src/util/constants.ts:8](https://github.com/paritytech/txwrapper/blob/0fb6503/src/util/constants.ts#L8)*
+*Defined in [src/util/constants.ts:8](https://github.com/paritytech/txwrapper/blob/fcbe6db/src/util/constants.ts#L8)*
 
 Prefix for SS58-encoded addresses on Polkadot
