--- docs/modules/_util_interfaces_.md
@@ -1,10 +1,0 @@
-[@amaurymartiny/txwrapper](../README.md) › [Globals](../globals.md) › ["util/interfaces"](_util_interfaces_.md)
-
-# External module: "util/interfaces"
-
-## Index
-
-### Interfaces
-
-* [BaseTxInfo](../interfaces/_util_interfaces_.basetxinfo.md)
-* [UnsignedTransaction](../interfaces/_util_interfaces_.unsignedtransaction.md)
