+++ docs/modules/_util_types_.md
@@ -1,0 +1,10 @@
+[@amaurymartiny/txwrapper](../README.md) › [Globals](../globals.md) › ["util/types"](_util_types_.md)
+
+# External module: "util/types"
+
+## Index
+
+### Interfaces
+
+* [BaseTxInfo](../interfaces/_util_types_.basetxinfo.md)
+* [UnsignedTransaction](../interfaces/_util_types_.unsignedtransaction.md)
