+++ src/decode/decodeUnsignedTx.ts
@@ -10,8 +10,7 @@ import { createType, Metadata, TypeRegistry } from '@polkadot/types';
 import { setSS58Format } from '@polkadot/util-crypto';
 
 import { BLOCKTIME, KUSAMA_SS58_FORMAT } from '../util/constants';
-import { BaseTxInfo } from '../util/interfaces';
-import { UnsignedTransaction } from '../util/interfaces';
+import { BaseTxInfo, UnsignedTransaction } from '../util/types';
 import { getMethodData } from './decodeUtils';
 
 export interface DecodedUnsignedTx extends BaseTxInfo {
